requirements.txt
asgiref==3.7.2
crispy-tailwind==0.5.0
Django==4.2.7
django-crispy-forms==2.1
django-filter==23.4
django-tailwind==3.6.0
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
